Version with embredd IFRAME.
Tried to embed an IFRAME to have the list of dictionary entries
scrollable and the "title" in fixed positiom. Alas, it does not seem to
work, plus the latest Firefox does not seem to understand IFRAMES with
'srcdoc'.  Also, the 'seamless' flag does not seem to be working
properly in Safari and Chrome.

So.  Back to a simpler design.