Ensured that packages and systems always have a doc-string.
If packages and systems didn’t have a doc-string they didn’t end up in
the list of items to document.  But systems and packages are the root
of the documents hierachy; thus, wit no doc-string, the documentation
would not have where to start from.  Adding a minimal doc-string was
the simplest way to fix the problem.